+++ extensions/advanced-replace/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Regex Replace Changelog
 
+## [Additional Replacement Method] - {PR_MERGE_DATE}
+
+- New type prop: directReplace | cutPaste
+- Direct Replace: Search & Replace directly in input string
+- Dropdown on create form changes fields accordingly
+
 ## [Initial Version] - 2025-01-12
 
 - Initial replacement method